refactor[#117] : 비즈니스 로직이 담긴 폴더명 model -> lib 으로 이름 변경
